test properties modified
Former-commit-id: a29b6bc7713827be9842b536e500ebb87cabbd69
Former-commit-id: 888307a698832fc2d4979f36288be1ebbadb4e86

+++ kernel/kernel-idgenerator-tsp/src/main/java/io/mosip/kernel/idgenerator/tsp/dto/TspResponseDTO.java
@@ -19,6 +19,6 @@ public class TspResponseDTO {
 	/**
 	 * The TSPID generated.
 	 */
-	private int tspId;
+	private long tspId;
 
 }
